# ggml-cuda : perform cublas mat mul of quantized types as f16