Improves prompt processing speed with quantized types with mmq disabled only (`-nommq`).

Essentially this is the same as https://github.com/ggerganov/llama.cpp/pull/3370, extended to quantized types by dequantizing to fp16. 


| model                          |       size |        mmq | test       |       master t/s |        PR t/s | speedup |
| ------------------------------ | ---------: | ---------: | ---------- | ---------------: | ------------: | ------: |
| llama 7B mostly Q2_K           |   2.63 GiB |          0 | pp 512     |   1740.05 ± 2.18 | 3422.52 ± 26.59 |    1.97 |
| llama 7B mostly Q3_K - Large   |   3.35 GiB |          0 | pp 512     |  1704.44 ± 36.02 | 3434.08 ± 21.88 |    2.02 |
| llama 7B mostly Q3_K - Medium  |   3.07 GiB |          0 | pp 512     |   1725.62 ± 2.40 | 3423.26 ± 44.50 |    1.98 |
| llama 7B mostly Q3_K - Small   |   2.75 GiB |          0 | pp 512     |  1720.28 ± 17.54 | 3415.61 ± 15.86 |    1.98 |
| llama 7B mostly Q4_0           |   3.56 GiB |          0 | pp 512     |   1705.19 ± 5.29 | 3230.66 ± 16.41 |    1.89 |
| llama 7B mostly Q4_1           |   3.95 GiB |          0 | pp 512     |  1696.79 ± 16.01 | 3241.20 ± 25.87 |    1.91 |
| llama 7B mostly Q4_K - Medium  |   3.80 GiB |          0 | pp 512     |  1718.43 ± 16.33 | 3507.96 ± 8.80  |    2.04 |
| llama 7B mostly Q4_K - Small   |   3.59 GiB |          0 | pp 512     |   1727.00 ± 4.03 | 3413.81 ± 97.65 |    1.98 |
| llama 7B mostly Q5_0           |   4.33 GiB |          0 | pp 512     |   1695.06 ± 6.91 | 3172.95 ± 15.14 |    1.87 |
| llama 7B mostly Q5_1           |   4.72 GiB |          0 | pp 512     |   1697.97 ± 4.07 | 3179.81 ± 35.89 |    1.87 |
| llama 7B mostly Q5_K - Medium  |   4.45 GiB |          0 | pp 512     |   1721.78 ± 2.12 | 3460.38 ± 34.02 |    2.01 |
| llama 7B mostly Q5_K - Small   |   4.33 GiB |          0 | pp 512     |   1722.66 ± 4.93 | 3474.46 ± 36.03 |    2.02 |
| llama 7B mostly Q6_K           |   5.15 GiB |          0 | pp 512     |   1712.02 ± 3.62 | 3468.20 ± 23.51 |    2.03 |
| llama 7B mostly Q8_0           |   6.67 GiB |          0 | pp 512     |   1685.94 ± 8.08 | 3176.80 ± 51.69 |    1.88 |


For comparison, this is the performance that I get with mmq enabled (the default):

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| llama 7B mostly Q2_K           |   2.63 GiB |     6.74 B | CUDA       |  99 | pp 512     |   1814.65 ± 4.73 |
| llama 7B mostly Q3_K - Large   |   3.35 GiB |     6.74 B | CUDA       |  99 | pp 512     |  1922.97 ± 19.27 |
| llama 7B mostly Q3_K - Medium  |   3.07 GiB |     6.74 B | CUDA       |  99 | pp 512     |   2009.16 ± 8.20 |
| llama 7B mostly Q3_K - Small   |   2.75 GiB |     6.74 B | CUDA       |  99 | pp 512     |  1901.53 ± 31.49 |
| llama 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CUDA       |  99 | pp 512     |   2420.71 ± 9.51 |
| llama 7B mostly Q4_1           |   3.95 GiB |     6.74 B | CUDA       |  99 | pp 512     |  2099.10 ± 31.81 |
| llama 7B mostly Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       |  99 | pp 512     |   2220.84 ± 1.78 |
| llama 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CUDA       |  99 | pp 512     |  2181.68 ± 54.97 |
| llama 7B mostly Q5_0           |   4.33 GiB |     6.74 B | CUDA       |  99 | pp 512     |   2191.70 ± 5.07 |
| llama 7B mostly Q5_1           |   4.72 GiB |     6.74 B | CUDA       |  99 | pp 512     |   1945.15 ± 6.50 |
| llama 7B mostly Q5_K - Medium  |   4.45 GiB |     6.74 B | CUDA       |  99 | pp 512     |   2070.01 ± 5.23 |
| llama 7B mostly Q5_K - Small   |   4.33 GiB |     6.74 B | CUDA       |  99 | pp 512     |  2044.41 ± 12.79 |
| llama 7B mostly Q6_K           |   5.15 GiB |     6.74 B | CUDA       |  99 | pp 512     |   2125.72 ± 7.12 |
| llama 7B mostly Q8_0           |   6.67 GiB |     6.74 B | CUDA       |  99 | pp 512     |  2346.21 ± 35.02 |