@slaren 2023-09-30
Perplexity looks good:
| model | ppl |
| --- | --- |
| 7B/ggml-model-f16.gguf       |  5.9073 +/- 0.03309 |
| 7B/ggml-model-Q2_K.gguf      |  6.5864 +/- 0.03755 |
| 7B/ggml-model-Q3_K_S.gguf    |  6.4524 +/- 0.03672 |
| 7B/ggml-model-Q3_K.gguf      |  6.1548 +/- 0.03456 |
| 7B/ggml-model-Q3_K_L.gguf    |  6.0866 +/- 0.03416 |
| 7B/ggml-model-Q4_0.gguf      |  6.1159 +/- 0.03504 |
| 7B/ggml-model-Q4_1.gguf      |  6.0655 +/- 0.03400 |
| 7B/ggml-model-Q4_K_S.gguf    |  6.0067 +/- 0.03371 |
| 7B/ggml-model-Q4_K.gguf      |  5.9616 +/- 0.03342 |
| 7B/ggml-model-Q5_0.gguf      |  5.9814 +/- 0.03409 |
| 7B/ggml-model-Q5_1.gguf      |  5.9418 +/- 0.03328 |
| 7B/ggml-model-Q5_K_S.gguf    |  5.9463 +/- 0.03330 |
| 7B/ggml-model-Q5_K.gguf      |  5.9196 +/- 0.03317 |
| 7B/ggml-model-Q6_K.gguf      |  5.9076 +/- 0.03309 |
| 7B/ggml-model-Q8_0.gguf      |  5.9078 +/- 0.03309 |