@Ph0rk0z 2023-09-30
Will this murder P40?