@Ph0rk0z 2023-09-30
Right but what happens if one gpu is pascal and one GPU is ampere? Will it go with the lowest cuda version for all?