@slaren 2023-09-30
This was already only used on the main GPU, but I think that even that may not work properly when converting `dst` to fp32 due to synchronization issues. So this is completely disabled with multi GPU now, the fp32 mat mul is used when using more than one GPU.