@ggerganov 2023-09-30
I update the A100 numbers using this PR: https://github.com/ggerganov/llama.cpp/discussions/3359