@Dampfinchen 2023-09-30
This increases VRAM for a reason. With this build and using --nommap my q4K_S model no longer fits in VRAM and it slows down dramatically.