@slaren 2023-09-30
> Hopefully this change can be made to work with mmq as well.

Once support for tensor cores is added to mmq, it will be as fast or faster than cublas again, while still using less VRAM. For now, cublas is the only way to use tensor cores.