@Dampfinchen 2023-09-30
> > Hopefully this change can be made to work with mmq as well.
> 
> Once support for tensor cores is added to mmq, it will be as fast or faster than cublas again, while still using less VRAM. For now, cublas is the only way to use tensor cores.

Just ran NSight and can confirm the tensor cores are, for the first time ever, being used to their full extent.

![tensorcores](https://github.com/ggerganov/llama.cpp/assets/59751859/3b0e3bcd-9cf7-4bac-a365-c0cb72964459)

Awesome work! Now fingers crossed its easy to enable tensor core support for mmq as well. If mmq (which was faster than cublas before this committ) can benefit from it as well, then we are definately in for another revolution here. Exciting stuff!