@YellowRoseCx 2023-09-30
Would be interesting to see how the changes affect AMD users