@JohannesGaessler 2023-10-05
>Once support for tensor cores is added to mmq, it will be as fast or faster than cublas again, while still using less VRAM. 

~2 weeks ago I did a prototype implementation for mmq using tensor cores and was **not** able to get better performance. From what I can tell a prerequisite to getting good tensor core utilization would be to [load data asynchronously](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#asynchronous-data-copies). As of right now the mmq compute pipeline utilization (without tensor cores) is only ~50%.